requirements.txt
Flask==1.0.2
pyyaml==3.13